requirements.txt
ccxt>=2.9.14
pandas>=1.5.3
numpy>=1.24.3
backtrader>=1.9.78
matplotlib>=3.7.1
python-dotenv>=1.0.0
websockets>=11.0.3
aiohttp>=3.8.4
flask>=2.3.2
openai>=1.3.7
SQLAlchemy>=2.0.19
slack-sdk>=3.21.3
asyncio>=3.4.3
# New dependencies
pytest>=7.3.1
pytest-cov>=4.1.0
streamlit>=1.26.0
plotly>=5.15.0
optuna>=3.2.0
hyperopt>=0.2.7
# EMA Crossover strategy dependencies
backtesting>=0.3.3
pandas>=1.3.0
numpy>=1.20.0
matplotlib>=3.4.0
seaborn>=0.11.0
streamlit>=1.10.0
plotly>=5.5.0
requests>=2.27.0
ccxt>=2.5.0 